Show only warning and error empty states - Inline MobileControls to Solver
