+++ packages/scrabble-solver/src/components/Solver/components/EmptyState/EmptyState.module.scss
@@ -5,8 +5,6 @@
   border: var(--border);
   border-radius: var(--border--radius);
   background-color: white;
-  box-shadow: var(--box-shadow);
-  cursor: pointer;
   text-align: center;
 }
