--- packages/scrabble-solver/src/components/Solver/components/MobileControls/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MobileControls';
