+++ packages/scrabble-solver/src/components/Solver/components/index.ts
@@ -1,5 +1,4 @@
 export { default as ApplyButton } from './ApplyButton';
 export { default as EmptyState } from './EmptyState';
-export { default as MobileControls } from './MobileControls';
 export { default as ResultCandidatePicker } from './ResultCandidatePicker';
 export { default as SolveButton } from './SolveButton';
